backend/requirements.txt
# Backend requirements for Ingredient Scanner Web App
# Run: pip install -r requirements.txt

# Core Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database & Authentication
supabase==2.0.2
postgrest==0.13.0
python-dotenv==1.0.0

# Machine Learning & Computer Vision (existing)
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.21.0
torch>=1.11.0
torchvision>=0.12.0
Pillow>=9.0.0

# API & Data Processing
requests==2.31.0
python-multipart==0.0.6
Werkzeug==3.0.1

# Utilities
python-dateutil==2.8.2
bcrypt==4.1.2
email-validator==2.1.0

# Development (optional)
pytest==7.4.3
pytest-flask==1.3.0